+++ template/utils/debounce.ts
@@ -2,7 +2,7 @@ export default function debounce(
   func: (...args: any[]) => void,
   delay: number
 ) {
-  let timeout: NodeJS.Timeout | null = null;
+  let timeout: ReturnType<typeof setTimeout> | null = null;
   return (...args: any[]) => {
     if (timeout) clearTimeout(timeout);
     timeout = setTimeout(() => {